% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/fake_libc_utils.R
\name{fake_libc_path}
\alias{fake_libc_path}
\title{Get the path to the installed fake_libc headers}
\usage{
fake_libc_path()
}
\value{
Character scalar with the path to fake_libc headers
}
\description{
Returns the absolute path to the inst/fake_libc directory in the installed package.
}
