% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_defines_from_file}
\alias{get_defines_from_file}
\title{This function will use the configured C compiler to list macro definitions (\code{-dM -E}) if
\code{use_cpp = TRUE} and a compiler is available; otherwise, a simple scan of \verb{#define} lines
is used as a fallback.}
\usage{
get_defines_from_file(file, use_cpp = TRUE, cc = r_cc(), ccflags = NULL)
}
\arguments{
\item{file}{Path to a header file}

\item{use_cpp}{Logical; use the C preprocessor if available}

\item{cc}{Compiler string; passed to \code{system2} if \code{use_cpp = TRUE}.}

\item{ccflags}{Additional flags for the compiler}
}
\value{
Character vector of macro names defined in \code{file}
}
\description{
This function will use the configured C compiler to list macro definitions (\code{-dM -E}) if
\code{use_cpp = TRUE} and a compiler is available; otherwise, a simple scan of \verb{#define} lines
is used as a fallback.
}
