% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_enum_members_from_root}
\alias{get_enum_members_from_root}
\title{Extract enum members from a parsed header}
\usage{
get_enum_members_from_root(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame with enum member names and values.
}
\description{
Extract enum members from a parsed header
}
