% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_enum_nodes}
\alias{get_enum_nodes}
\title{Extract enum names from a parsed header}
\usage{
get_enum_nodes(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame with enum names.
}
\description{
Extract enum names from a parsed header
}
