% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_function_nodes}
\alias{get_function_nodes}
\title{Extract function names (declarations and definitions) from a root}
\usage{
get_function_nodes(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame with function captures.
}
\description{
Returns a data frame with \code{capture_name}, \code{text}, \code{start_line},
and \code{start_col}.
}
