% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_globals_from_root}
\alias{get_globals_from_root}
\title{Extract global variable names from a parsed tree root}
\usage{
get_globals_from_root(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame with top-level global names.
}
\description{
Extract global variable names from a parsed tree root
}
