% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_globals_with_types_from_root}
\alias{get_globals_with_types_from_root}
\title{Extract global variables with types from a parsed tree root}
\usage{
get_globals_with_types_from_root(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame with global names and C types.
}
\description{
Extract global variables with types from a parsed tree root
}
