% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_struct_members}
\alias{get_struct_members}
\title{Extract members of structs (including nested anonymous struct members)}
\usage{
get_struct_members(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame describing struct members, including bitfields.
}
\description{
Extract members of structs (including nested anonymous struct members)
}
