% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_union_members_from_root}
\alias{get_union_members_from_root}
\title{Extract members of unions}
\usage{
get_union_members_from_root(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame describing union members, including bitfields.
}
\description{
Extract members of unions
}
