% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{get_union_nodes}
\alias{get_union_nodes}
\title{Extract union names from a parsed header}
\usage{
get_union_nodes(root)
}
\arguments{
\item{root}{A tree-sitter root node.}
}
\value{
Data frame with union names.
}
\description{
Extract union names from a parsed header
}
