% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/parse-headers.R
\name{r_cc}
\alias{r_cc}
\title{Return the default R-configured C compiler (possibly with flags)}
\usage{
r_cc()
}
\value{
Character scalar with compiler program (or empty string).
}
\description{
This function queries common places to find the C compiler used by R.
It checks Sys.getenv("CC"), then \verb{R CMD config CC}, and finally \code{cc} on PATH.
The returned value may include flags (e.g., 'gcc -std=...').
}
