% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/zzz.R
\docType{package}
\name{treesitter.c-package}
\alias{treesitter.c}
\alias{treesitter.c-package}
\title{treesitter.c: 'R' bindings to the 'C' Grammar for Tree-Sitter}
\description{
Provides bindings to a 'C' grammar for Tree-sitter, to be used alongside the 'treesitter' package. Tree-sitter builds concrete syntax trees for source files and can efficiently update them as files are edited.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://sounkou-bioinfo.github.io/treesitter.c/}
}

}
\author{
\strong{Maintainer}: Sounkou Mahamane Mahamane \email{stoure@example.com}

Other contributors:
\itemize{
  \item Tree-sitter authors (Tree-sitter C grammar) [copyright holder]
}

}
\keyword{internal}
